pyproject.toml
[project]
name = "moomoo-mcp-server"
version = "0.1.0"
description = "MCP server for Moomoo OpenD"
dependencies = [
    "mcp[cli]",
    "futu-api",
    "python-dotenv",
    "pydantic"
]
requires-python = ">=3.10"

[project.scripts]
moomoo-mcp = "moomoo_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
